guix: convert to Guile 3
Also with some additional cleanup such as adding Guile module paths for
covering a lot of the Guile modules especially coming from third-party
package repos (e.g., nixpkgs) and updating the Guile modules to use
Guile 3.x instead of overriding them to use it. It's not cleaner nor
more elegant but it reassuring and we can still force all packages to
use it if nothing else.