shells: remove non-devt shells
While the `guix` devshell is mainly for easily building the
package in Guix when navigating the source code, it does have some use
case for wanting to easily provision a dev't environment for the
codebase. However, it doesn't make sense at this point since
contributing requires Guix to be installed in the first place and it
also has some setup involved.