guile-lzlib: use lzlib with shared objects
It apparently fails due to the lack of dynamic library files from lzlib.
It seems like it cannot make use of the archived object file from the
compiled package.

This is what most likely makes Guix unable to use substitutes since most
of them from the cache are compressed with lzip.